+++ src/main/kotlin/com/onsafe/backend/domain/logs/model/entity/FallLog.kt
@@ -9,5 +9,6 @@ data class FallLog(
     val score: Float,
     val fall: Boolean,
     val isConfirmed: Boolean = false,
+    val imageUrl: String? = null,
     val timestamp: LocalDateTime = LocalDateTime.now()
 )